# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Buildspace API",
    "version": "0.1.0",
    "description": "Separate standalone backend contract. Full request schemas live alongside Hono routes in src/modules."
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
...
  "paths": {
...
    "/api/v1/spaces/{slug}/planning-notes": {
      "get": {
        "summary": "List planning-notes",
        "tags": [
          "Workspace"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      },
      "post": {
        "summary": "Create planning-notes",
        "tags": [
          "Workspace"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
    "/api/v1/spaces/{slug}/roadmap-items": {
      "get": {
        "summary": "List roadmap-items",
        "tags": [
          "Workspace"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      },
      "post": {
        "summary": "Create roadmap-items",
        "tags": [
          "Workspace"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
...
    "/api/v1/spaces/{slug}/planning-notes/{noteId}/promote": {
      "post": {
        "summary": "Promote note to task",
        "tags": [
          "Workspace"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "summary": "Notifications",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
    "/api/v1/notifications/{notificationId}/read": {
      "post": {
        "summary": "Mark notification read",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
    "/api/v1/notifications/read-all": {
      "post": {
        "summary": "Mark all notifications read",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
...
    "/api/v1/spaces/{slug}/integrations": {
      "get": {
        "summary": "Read integrations",
        "tags": [
          "Realtime and AI"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      },
      "post": {
        "summary": "Write integrations",
        "tags": [
          "Realtime and AI"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
    "/api/v1/spaces/{slug}/design-links": {
      "get": {
        "summary": "Read design-links",
        "tags": [
          "Realtime and AI"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      },
      "post": {
        "summary": "Write design-links",
        "tags": [
          "Realtime and AI"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
...
    "/api/v1/integrations/github/webhook": {
      "post": {
        "summary": "GitHub signed webhook receiver",
        "tags": [
          "Integrations"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "buildspace_session"
      }
    }
  }
}